server/requirements.txt
numpy>=1.26.0
requests
pyedflib>=0.1.36
mne>=1.6.0
fastapi>=0.68.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
redis>=5.0.1
aiofiles>=23.2.1
uvicorn>=0.15.0
loguru>=0.5.3
gunicorn
cryptography>=41.0.0
email-validator>=2.0.0
pyhumps>=3.8.0
asyncpg>=0.30.0
click>=8.1.8
minio>=7.2.15
prometheus-client>=0.21.1
dda-py>=0.1.1

# GraphQL dependencies
strawberry-graphql>=0.96.0
gql>=3.4.1
requests>=2.31.0
requests-toolbelt>=1.0.0

sqlalchemy>=1.4.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
aiohttp>=3.9.3
psycopg2-binary>=2.9.5  # PostgreSQL adapter

# Testing dependencies
tox>=4.0.0
pre-commit>=3.5.0
ruff>=0.1.6
pytest>=7.0.0
httpx>=0.24.0
requests_mock>=1.11.0
